MBarrier change split-out from #990 (#1070)
Changes:
- Added index lowering for `MBarrierInit` and `MBarrierInvalidate`
- Indexing: In the lowered index, just use the index itself, don't
generate a `kir::TensorIndex` object.
- Added a new IR node `MBarrierArriveExpectTx`, which can be used in
combination with `cp.async` and `cp.async.bulk` to let the barrier wait
for completion of such operations.
- `alias_memory.cpp`: mark `MBarrierInit` as the first write of memory
barrier, and `MBarrierInvalidate` as the last read. This way, the smem
reuse pass can correctly reuse and assign addresses of the mbarrier. For
this to work, `MBarrierInit` and `MBarrierInvalidate` must be inserted
before the memory reuse pass, but arrive and wait can be inserted later.

The changes in this PR is hard to test without TMA load, so no test is
added in this PR. Testing of this PR will be done in #990.