Cleanup after Nvfuser removal from pytorch (#1009)
This PR patches the build of nvfuser project to make sure it works after
the nvfuser removal from pytorch. Installations using either `python
setup.py develop` or a _pre-built pip wheel_ is covered.

_Using `python setup.py install` is not supported._

This PR patches the RPATH of `libnvfuser.co` and `libnvfuser_codegen.so`
to let them find `libcudart.so` from the relative path that are
installed from pypi.