Improve and generalize codegen diff tool (#1030)
This change:
- Moves codegen diff tools to `tools/codediff`.
- Creates a `run_command.sh` tool for collecting kernel information from
arbitrary commands.
- Add parser for `nvfuser_bench` and any other google benchmark binary.
This is automatically detected by `run_command.sh` if `nvfuser_bench` is
part of the command line, but can be overridden with `run_command.sh -t
<command type>`.
- Add parser for pytest tests. This is automatically detected if
`python_tests` or `pytest` is in the command line, and can be overridden
with `-t`. **Note that pytest must be run with the `-v` and `-s`
options, or their equivalents**.
- Renamed the diff tool to `diff_report.py` and takes pairs of output
directories from the `run_command.sh` tool.
- Adds json output from `diff_report.py`.
- Collects and displays PTX alongside CUDA diffs.
- Collects and displays LaunchParams, which lets us see the block, grid,
and dynamic shared memory used by each kernel launch. If more than one
set of LaunchParams are detected for a single kernel, only the first set
is shown. Parsing LaunchParams currently does not work very well in
pytest tests, as the association between launch params and kernels is
easily lost.
- Handles google benchmark output and pytest output, grouping each by
test/benchmark in the HTML output.
- Shows benchmark GPU time in the diff.
- Improves misc. aspects of the HTML output.