Add draft of sequential gridreduce functions to runtime
There are three pieces that need to be generated:
- initial sync
- reduce element
- final sync

The next parts will be tricky. We need to represent _where_ to place the
syncs, and where to allocate the temporary buffer. One option is to
always place them at the top scope which is the simplest. If we place
them within a loop, we need to treat the reduction as if it is
persistent, which just means we reset the semaphore in the last block
during the final sync. Note that we will need an index into this buffer
as well, and that cannot necessarily be equal to the index of the
output, since that might be inlined further than the temporary buffer.
Finally, as a separate PR some day, we could try to alias the temporary
buffer to an output.