Record GPU properties as context in nvfuser_bench
We currently record GPU name as the "GPU" field in additional context.
This means it is printed when the benchmark runs, and it appears as an
item in the "context" section when using `--benchmark_out_format=json`.
This change simply extracts more fields and populates the context in the
same way. Particularly useful fields are the SM and Memory clock rates
and memory limits. You can also compute peak bandwidth for the device
from the memory clock rate and bus width fields (from
`kernel_cache.cpp`):
```
// Peak bandwidth calculation:
// Bus width is given in bits, so dividing by 8 converts to bytes.
// Clock is given in kHz. 1 GB = 1e9 bytes (don't report GiB = 1024^3 bytes)
// A factor of 2 is multiplied to account for double data rate (DDR):
// (clock in kHz * width in bits) * (1000 Hz / kHz) * (1 GB / 8e9 bits) * 2
// factor = 2.5e-7
```

The GPU name is now called "gpu_name", and other GPU properties start
with "gpu_*". To keep casing consistent with the default fields, "Host"
is renamed to "host".