Make sure stdout capture is stopped at the end of each test (#1143)
Fixes #1142 

To make it easy to understand why this fix is necessary, I temporarily
added two new tests just to expose the issue.

```
 ./bin/nvfuser_tests --gtest_filter="*CaptureTest*"
```

This should run `CaptureTest1` and then `CaptureTest2`. The former fails
before stopping the capture, so with the current TOT, the issue of #1142
should happen. This PR makes sure the capture is stopped no matter if
the test itself succeeds.