Set the allocation domain of mma output as leaf domain and remove `accumulator_stride` (#1210)
Matmul support was initially developed before we invented the concept
"allocation domain". Without the concept of "allocation domain", our
codegen has lots of limitations. One limitation is, the allocation is
always based on the order of the rFactor domain, even for tensors
allocated on shared memory or local memory. For example, if I have a
tenor `[M, N]`, and I do some splits and reorder to get `[BIDx{M/2},
TIDx{N/3/5}, 3, 2, 5]` and the memory layout of the register will always
be `[2, 3, 5]`, regardless of how we reorder these three numbers in leaf
domain.

Due to this limitation, scheduling the layout of MMA output was
fundamentally impossible. For example, in the [memory layout for Volta
MMA](https://docs.nvidia.com/cuda/parallel-thread-execution/index.html#matrix-fragments-for-mma-m8n8k4-with-f16-floating-point-type):

![](https://docs.nvidia.com/cuda/parallel-thread-execution/_images/mma-884-C-f32-1.png)
the memory layout of register should be:
```
N --(splits)--> [N/8, 2*, 2**, 2***]
M --(splits)--> [M/4, 2', 2'']
register layout: [2*, 2', 2***]
```
which has interleaved M and N.

The workaround? If you look at the code in `tensorcore.cu`, each MMA op
has a template argument `acc_stride`. For example, in this function:
```CUDA
template <int acc_stride>
DEVICE_INLINE void initM16N8K16TN(Array<float, 4, 4>* accumulator) {
  float* _C = reinterpret_cast<float*>(accumulator);
  _C[0] = 0;
  _C[1] = 0;
  _C[acc_stride] = 0;
  _C[acc_stride + 1] = 0;
}
```
we accept an array[4] as a parameter, however, we are intentionally
doing out of boundary access inside this function, instead of using the
four items passed as arguments. This approach in practice works, but is
definitely not the correct way of doing engineering.

In the original design, there would be a "transpose swizzle" which takes
care of this, but that never got implemented. Today, we have a better
solution: we can just set the allocation domain to the desired order we
want, and the `accumulator` passed in will automatically contain the
elements we want. The `accumulator_stride` hack is no longer needed.

The major change of this PR is actually just three lines, see the change
in `csrc/tensor_view.cpp::TensorView::applyMmaSwizzle`, the rest of this
PR are just clean ups. These three lines just set the allocation domain
to leaf domain, which has the correct memory layout.

cc: @jjsjann123 @wujingyue who may be interested in this PR because this
is the first time that allocation domain get used in a scheduler to do
some serious work. Feel free to ask questions or leave comments.
cc: @naoyam as always