Runtime Support for new parallel type (#1022)
# What's in this patch:
- support Fusion with tensor which are sharded across devices (with the
new parallel type DiDx), with the following restrictions:
- restrictions described in #918, basically, tensors can only be shared
on their first dimension, on 1D device meshes. This implies that tensors
are either fully sharded or fully replicated
- For now, the Pipeline has to be set manually, in a way that there is
no communication besides the inter-stages communications.
  - The only allowed Expr at inter-stage is `Set`
- We add testing utilities to compare the results of a Fusion executed
on a single device with the multi-device version. The tests are
organized to be modular, minimize code duplication, and improve
performance



# Implementation details and motivations:
- **Runtime:**
- **Use `lowerCommunication` in the executor**. See the function `void
PipelineExecutor::handle(PipelineCommunication* c)` in
[multidevice/executor.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-d783a79fdc189059065a81ba9acb6d6c3d3ecc5c70188706febc97d7fea5aa36).
We remove the old implementation, and instead, we call the function
`lowerCommunication` (introduced recently in #918) that transforms the
`PipelineCommunication` into a series of `Communication`. Then we `post`
and `wait` these `Communications`
  - **Add Option to disable auto-scheduling**.
- We add in `PipelineStage` an attribute `bool auto_schedule` (see
[multidevice/pipeline.h](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-e66fbfbcfc303fec7ffb2c82e1f130dd7b72039a5601f8a146e2a0681824c725)).
Before, we always used auto-scheduling for each different stage -- but
now we want to manually set the parallel type and so we want to be able
to turn off the auto-scheduling.
- We thus handle this bool in the function `void
PipelineExecutor::handle(PipelineStage* c)` in
[multidevice/executor.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-d783a79fdc189059065a81ba9acb6d6c3d3ecc5c70188706febc97d7fea5aa36),
where we either use `FusionExecutor` or `FusionExecutorCache` depending
on `bool auto_schedule`.
- We need to implement `FusionExecutor::allocOutputSpace`, which is done
in
[csrc/executor.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-e72d395ea8a3ab3d2a909dfb3677beac1bbd89336b54ee858867f652b48be4f0)
- **error handling in MultiDeviceRuntime:** small change in
`MultiDeviceRuntime::validate()`, inside
[multidevice/runtime.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-476e4d39a0bfe025d0030a81ee4a99d7d39c950fa096d109e2b4e2ee89d861fb).
Before, this function was asserting, now it returns a string with an
error msg. Empty string means no error. This is useful so that in the
test we can handle the error as we want, notably we may want to simply
skip the test if there are not enough ranks, instead of throwing an
error.
- **Test:**
- **Share the `Communicator` across all tests:** this is needed for two
reasons: 1) creating the `Communicator` is a very costly operation, and
2) we cannot create multiple `Communicators` on the same `address:port`
(this restriction comes from pytorch's ProcessGroups). We point out that
defining a global static Communicator is not valid and often leads toa
segfault during cleanup. Instead, To achieve this, we define the class
`class MultiDeviceEnvironment : public testing::Environment` and add an
instance to the GTEST environment with the command
`testing::AddGlobalTestEnvironment(new MultiDeviceEnvironment))`. See
[test/multidevice.h](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-0d4c54d65e1463bf971ea74222f9653194a9fff61015202ed105b01883d76b2b)
and
[test/multidevice.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-00a0e9da3365a4a923933dd9cf701175f28c5c8f157741884d5242ad0a02a49d)
- **Tests utilities and structure:** we define different classes and
methods in
[test/multidevice.h](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-0d4c54d65e1463bf971ea74222f9653194a9fff61015202ed105b01883d76b2b).
The structure we provide allow to avoid a lot of code duplication in the
tests, and to make them more adaptable with future developments. Among
the utilities we provide, `PipelineTest::validate()` allows to run a
pipeline on a multidevice setting and to compare the results with the
output of the same fusion executed on a single device.
- **New tests**: We provide new tests using pipelines and the new
parallel types in
[test/test_multidevice_pipeline.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-ccc7195e288586558c5f7aff6761ba0f5825de3ce39e469b0858e86b3fa737b1)
(this file has been renamed and was previously named
[test/test_multidevice.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-9c5f8ee781e2c6ee3eb63e8f628becfab0c0514d5e0a21266811a76eabfda254))
- **Clean the implementations of the old tests** in
[test/test_multidevice_communications.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-b8f1ce82763ea4f5fb61b4773a1f4232e1325fa63b63616c7ffee7c0a48b2296)`
by leveraging the new classes from
[test/multidevice.h](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-0d4c54d65e1463bf971ea74222f9653194a9fff61015202ed105b01883d76b2b)
- Remove a now obsolete check in
[test/main.cpp](https://github.com/NVIDIA/Fuser/pull/1022/files#diff-7afe085f03fa2399a9d1d962f6aaa46c84e72300d791c4aa657f56cc70f638e0).