Fix dangling dead uses in Fusion::removeVal (#1280)
Fixes #1270.

When a `Val` is removed from the `Fusion`, we currently first loop over
`val->uses()` and remove them all from the Fusion. This seems correct at
first glance, but it is incomplete since `val->uses()` actually only
gives all **live** uses. When there is dead code in the `Fusion` that
includes some uses of a val that is to be removed, we can wind up with
an expression that holds an invalid pointer to the removed value in its
`inputs()`. In #1270 this caused a segfault when the fusion was cloned
since that will clone not only live objects but also these dangerous
dangling dead ones.

This PR iterates over `exprs_` in `Fusion::removeVal()` in order to
definitively find uses, dead or alive. Note that we still do not check
attributes for the presence of an `attributeVal` equal to `val` so we
could probably still contrive a case that would lead to dangling
references. If we do check such attributes we should maybe produce an
error instead of silenlty removing the `Expr`. I am not sure how
prevalent `Val` attributes are or what would be preferred in those
cases.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>