Simplify index generation of ldmatrix and mma ops by allowing `kir::TensorIndex` to view the memory differently,  (#1263)
The dtype of `kir::TensorIndex` is no longer required to be the same as
its `view()`. If they are different, then a `reinterpret_cast` is
generated to convert the type. `ldmatrix` and the output of mma is
changed to utilize this feature. This PR pave the way towards lowering
`MmaOp` as `kir::Asm`.

This PR only adjusts the style of generated code, there should be no
behavioral change.