IdModel: Reproduce exact map (#1168)
This is the first in a series of PRs for the new IdModel infrastructure.
Extracted from the original PR (#32).

As the first PR, I tried to make it as small as possible while still
doing something non-trivial. In particular, all it does for now is to
construct the exact graph of domains, which corresponds to the exact map
in our current infrastructure. The `ComputeAtMap` class is meant to be
replaced by `IdModel`, which at this moment only has the exact graph but
will be added with the Permissive and Loop graphs in follow-up PRs.

To exercise the new graph, I modified the lowering to build the graph
when enabled with `NVFUSER_ENABLE=id_model`. It currently just build the
graph and does some validaiton.

The new exact graph and the current exact map have several differences.
One major difference is that the domain relationships are represented as
a DAG instead of just groupings. Graph nodes include groups of domains
as well as groups of domain expressions, and those groups are connected
with input-output dependencies. Much like we can traverse from a leaf
domain to its root domains, we can now traverse from a group of domains
to its defining expression groups and their inputs.

Aside from the graph-vs-grouping difference, the resulting groupings are
mostly the same, which is validated by comparing the two groupings.
However, there are two subtle differences between the two:

1. Swizzles are just skipped in the current exact map
(https://github.com/NVIDIA/Fuser/blob/main/csrc/compute_at_map.cpp#L476-L477),
whereas that's not the case in the IdModel exact graph.
2. The underlying graph structure allows for discovering more domains to
group together. For example, when two domains are exactly mapped and
both are split with the same factor, the split output domains are not
currently grouped together, whereas they are indeed mapped with this PR
thanks to the graph-based structure.

When comparing the new exact grouping with the current exact grouping in
`checkExactGraphEquivalence`
(https://github.com/NVIDIA/Fuser/pull/1168/files#diff-8e80d015bc40509fdd3a208da9c1166a2d85b49a924fab40cec21758143b3e36R32),
I extended the current exact map to account for the second difference.
The swizzle difference seems difficult to handle, so I just gave up
comparing them when a swizzle is used.

No existing functionality should be affected with this PR even with
`NVFUSER_ENABLE=id_model`, although generated code would be different as
we potentially create new domains, affecting variable names. By default,
the option is still turned off.

The plan after this PR would look like:

1. Enable the permissive graph
2. Enable the loop graph
3. Enable the loop promotion

At this point, the bulk of the original PR (#32) would be completed.
Things to do after them would include (in no particular order):

- Refactor some of the broadcast analyses with the loop promotion map,
e.g., the sync insertion
- Replace the current ComputeAtMap with IdModel
- Build the index graph
- Refactor the vectorization analysis

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>