Automatically load and store cache hierarchy when constructing FusionCache (#1229)
This PR enables loading and storing the fusion cache hierarchy
automatically from a common workspace.

When constructing the `FusionCache`, we deserialize from
`tmp/nvfuser_kernel_db/nvf_serde`. This step can be disabled by passing
a bool flag to `FusionCache::get`.

In python, we register with `atexit` the `FusionCache::serialize`
function to automatically serialize when exiting the program.

```python
# Use atexit to automatically call serialize on program exit
import atexit
from nvfuser import serialize as nv_serialize
atexit.register(nv_serialize)
```

Serialization saves the `FusionCache` to a per-process temporary file
first to avoid conflicts with multiple processes. Then, the temporary
file is renamed to the actual file. If the actual file already exists,
then the `std::filesystem::rename` call may fail or replace the actual
file. This approach is taken from PyTorch Jiterator.

The saved workspace binary may become out-of-date with the NvFuser
repository. If the binary cannot be verified, it is deleted and the
`FusionCache` is not restored.
 
Reference:
https://github.com/pytorch/pytorch/blob/main/aten/src/ATen/native/cuda/jit_utils.cpp#L1602-L1629