Fix invisible uses in DeadCodeRemover traversal (#1272)
This fixes #1246 which was a bug in `DeadCodeRemover`, which is used in
the `EmptyTensorRemover` pre-segmentation pass. In the fusion from that
issue, `T2` is erroneously removed from the Fusion, causing the Fusion
to appear empty. That removal occurs after the backward traversal,
during which it is marked dead despite the fact that it was earlier used
to define the new tensor `T6 = set(T2)` which was also earlier
registered to replace `T5`.

The underlying issue was that `Val::uses()` only returns uses that lead
to Fusion outputs (see `Fusion::resetTvUses()`). It will not just return
all the `Expr`s for which that `Val` is an input. So during the backward
traversal, even though `T2` should've been kept alive by `T6`'s
definition, that use was not returned so `allUsesDead()` returned
`true`.

This PR tracks future live uses of `Val`s when they are marked as live
in order to avoid this overzealous pruning.