Update mma unit tests (#1311)
`FusionAmpereMMA*` tests are intended to be a unit test specifically for
testing if the mma op is working. It makes no sense to have epilogue for
it. The epilogues have already been covered by other tests well, so
removing these tests.

Also, add composed mma tests for `m16n16k16`.