Unify `IterDomain::split` with `SplitTransform` and similar for merge (#1329)
Currently, `MergeTransform` and `SplitTransform`, which are the classes
used to perform reshape operations, do not re-use the splitting and
merging operations used in scheduling, namely the static methods
`IterDomain::{split,merge}`. This leads to needing to maintain changes
in both as I recently found out when debugging #1174. We can actually
already notice they have diverged with respect to
`IndexType::GatherScatter` handling.

This PR simply calls the static `IterDomain` method from each of the
reshape `*Transform` classes. Since those classes also need to set the
rfactor-product flag on the produced `IterDomain`s, I added a flag to
the static methods that is only used during reshape.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>