Encode mma macro enum as `uint64_t` and define hopper macros (#1362)
`MacroType` is a enum that specifies which mma instruction to use. In
the past, the underlying type for this enum is just the default given by
C++. This PR changes this to `uint64_t`, and explicitly specify a value
for each macro, so that the first two bits represent arch, the second
two bits represent `m`. This makes it possible to pull the information
of an enum without using a `switch`. I am trying to avoid `switch`
because hopper have a lot of macros, and having to write all of them
down is so boring...

This PR also adds hopper macros, although they are not implemented yet.