Make mma and ldmatrix unit tests really unit (#1375)
Currently, our mma "unit" tests uses ldmatrix to load data to registers.
This makes the "unit" test not really unit, because if we scheduled both
ldmatrix and mma wrongly, as long as they are wrong in a consistent way,
we can still get the correct result. Before
https://github.com/NVIDIA/Fuser/pull/1303, we were indeed intentionally
scheduling it wrongly but consistently to workaround the lack of
allocation domain. But today, we should have everything we need to
schedule it correctly, so in the unit test of mma, we should no longer
use `ldmatrix` to read data to registers. Instead, we should read from
gmem to registers directly, so that we can test if the memory format we
scheduled is correct.

On the other hand, we need to add a unit test for ldmatrix. Similarly,
we should not validate its results by using it with mma op. Instead, we
should use `ldmatrix` to do a copy kernel, there should be no matmul
involved.

This PR also cleans up some deprecated code.

---------

Co-authored-by: Andrzej Bekas <118676880+drzejan2@users.noreply.github.com>