Stride order broadcast update (#1370)
Fixes #1369 

Updated stride_order computation to preserve the semantic position for
implicit broadcast dimension, i.e. dimension with stride == 0.
This avoids unnecessary alloc_domain for expanded dimension, which would
temporarily avoid perf regression due to the lack of alloc_domain
support in many of our schedulers.

This is not a permanent solution. We would still want to plumb through
alloc_domain support on schedulers as well as vectorization analysis.
Nevertheless, the updated logic is still an improvement.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>