Support boolean inputs for `kir::Asm` and lower `cp.async` as `kir::Asm` (#1364)
Inline PTX does not have a builtin constraint for boolean values, we
have to convert the boolean to int first, and use `setp.ne.b32` to put
this int to predicate register. With this new feature added, we can now
lower `cp.async` as `kir::Asm`. This feature is also needed for
supporting Hopper MMA.