Initial support for Hopper MMA instructions (#1409)
This PR adds support for Hopper MMA instructions. Current limitations
are:
1. Each smem tensor must be exactly one tile.
2. Synchronization are hard coded before and after the MMA instructions,
therefore it is indeed synchronous.
3. Only support swizzle mode "none", which is pretty different from
other swizzle modes, but conceptually a bit simpler to start with.

In terms of testing, two new test suites are added: `MmaTest/Hopper.RS`
(where A is on register and B on smem) and `MmaTest/Hopper.RS` (where
both A and B are on smem). These two test suites will be instantiated
for all Hopper mma instructions, for all layouts, for all swizzle modes,
and for all dtypes (half and bf16), so there will be a great number of
new tests added, but I believe this exhaustive test is necessary.

For review, I recommend briefly read through [PTX
doc](https://docs.nvidia.com/cuda/parallel-thread-execution/index.html#asynchronous-warpgroup-level-matrix-multiply-accumulate-instructions)
first before reading this PR.

Also, be warned that, the smem memory format is not straightforward at
all (actually less straigntforward than register memory format in my
opinion). Especially, I was originally expecting that the instruction
tile for the no-swizzle mode will be just a simple row major or column
major matrix, but this seems to be not the case. For the case of
no-swizzle mode, you need to divide the instruction tile into 8x8 core
matrices, and put each core matrix into contiguous memory, which is
something like:
```C++
[M, K] -> [M/8, 8, K/8, 8] -> [M/8, K/8, 8, 8]
```
Also, another surprise to me is, the `imm-trans-A` and `imm-trans-B` is
actually ignored by the `wgmma` instruction for no-swizzle mode. So
regardless of your input layout, in smem, you need to manually transpose
it to the K-major layout (`NT`, `[K, M] x [K, N]`) in smem before
calling the `wgmma` instruction.

This PR is only initial support for Hopper MMA. Remaining things are:
- Support multi-tile
- Support other swizzle mode
- Be asynchronous