Avoid binding expanded extents when passed unexpanded tensor
This fixes a subtle bug that was exposed by #973. In the test
ExpandedBroadcastGlobalIntermediateTest.TheTest_CUDA, which currently
passes, we are binding an unexpanded output tensor of size (2, 1, 2)
even though that output should have an expanded size of `1 << 60 =
1152921504606846976`. The expression evaluator currently doesn't
complain that we are binding the value 1 to a constant int with value
1152921504606846976 since in fact the expanded extent of that axis has a
cast, i.e. it is the non-immediate constant integer
`(nvfuser_index_t)1152921504606846976`. That is the actual `Val*` being
bound, so the `ExpressionEvaluator` does not complain that the values
conflict, but once the cast is simplified away, we get an error.

This PR fixes this by avoiding binding to expanded axis if a simple
broadcast axis is passed in. However, whenever the input tensor has a
stride of zero, we do bind the extent even if the size is 1.