Unify commit-wait style sync (#1476)
`kir::CpAsyncWait` and `kir::CpAsyncCommit` is now renamed as
`kir::AsyncWait` and `kir::AsyncCommit`, and they are now also used for
synchronizing `cp.async.bulk` and `wgmma`. Using a unified IR node
reduces the number of code, and therefore in the future it is easier to
add another type of sync.