Relax the allocation domain constraint.  (#1492)
Currently, `MarkAliasPass` skips an output when its allocation domain is
set at all. This PR makes it smarter to check compliance.

This PR prepares #1478 for landing. Before segmentation, we'll run alias
analysis on the un-segmented fusion and mark layouts for aliasing.
During scheduling, we'll run alias analysis again on segmented fusions
and have to recognize the preferred layout is compliant with the
previously marked layout.