Duplicate the code of `Swizzle2D` and name it as `Swizzle` (#1493)
The current implementation of `Swizzle2D` is so weird that I don't want
to use it for implementing the smem layout of Hopper. So I selectively
copy-pasted its code and name the new op `Swizzle`. In the future, I am
expecting to deprecate `Swizzle2D`. Except for the newly added test,
this entire PR is copy-pasted from existing `Swizzle2D`. But not all of
`Swizzle2D` are duplicated. I want the new `Swizzle` op to remain simple
and stupid, that is, it will behave similar to `Split` and `Merge`:
- It is not handled separately in indexing, `IndexSwizzle` only applies
to `Swizzle2D` and is unrelated to the new `Swizzle`. The new `Swizzle`
is just indexed by the most common back propagation like `Split` and
`Merge`.
- In `TransformReplay` and `BestEffortReplay`, there is no mechanism for
skipping it
- Unlike `Swizzle2D`, the inputs and outputs of `Swizzle` are not mapped
in compute at map

And  I want the new `Swizzle` to remain like this in the future.

Currently, this new `Swizzle` is not capable of replacing the existing
`Swizzle2D` for supporting Turing and Ampere matmul, because the lack of
the above mentioned mechanism that I intentionally kept out. In the long
future, `Swizzle` in combination with allocation domain will be used to
support Turing and Ampere, in a way as described in
https://github.com/NVIDIA/Fuser/issues/986, but this requires us to be
able to do forward propagation for ID ops for indexing. In the near
future, `Swizzle` will be used to support swizzling smem for Hopper, at
least for supporting tests in
https://github.com/NVIDIA/Fuser/blob/main/test/test_mma.cpp.

I understand that, in general, code reuse should be preferred compared
to copy-pasting code, but for this specific case, I prefer to just copy
past the code. Because the `Swizzle2D` will be removed completely, and
in the intermediate state today, I do not want to share any code with
it, so that I will not have to validate if my changes in `Swizzle` is
compatible with the legacy `Swizzle2D`. I want to keep `Swizzle2D`
completely untouched, until one day I completely remove it. I believe
copy-pasting has the lowest maintenance cost.